Include redirect_map helper in the generic table definition
This change is required to use the redirect_map helper in the devmap table.
In Polycube, the table has to be shared across multiple eBPF programs but this
is not possible in the standard version of BCC.
This is a sort of workaround to allow shared devmaps to use the redirect_map helper
without having to modify the devmap table definition.

Signed-off-by: Sebastiano Miano <mianosebastiano@gmail.com>